refactor: 도메인 패키지 내부 분리
