+++ src/main/java/store/domain/product/ProductStorage.java
@@ -1,4 +1,4 @@
-package store.domain;
+package store.domain.product;
 
 import java.util.List;
 import java.util.Map;
