+++ src/main/java/store/domain/promotion/Promotions.java
@@ -1,4 +1,4 @@
-package store.domain;
+package store.domain.promotion;
 
 import java.util.List;
