+++ src/test/java/store/domain/PromotionFactoryTest.java
@@ -3,6 +3,7 @@ package store.domain;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import store.domain.promotion.PromotionFactory;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
