+++ src/test/java/store/service/ProductServiceTest.java
@@ -2,10 +2,10 @@ package store.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import store.domain.ProductStorage;
-import store.domain.Promotion;
-import store.domain.PromotionFactory;
-import store.domain.Promotions;
+import store.domain.product.ProductStorage;
+import store.domain.promotion.Promotion;
+import store.domain.promotion.PromotionFactory;
+import store.domain.promotion.Promotions;
 
 import java.util.List;
